Внесены изменения в роль менеджер и юзерменеджмент сервис

+++ PizzaWebAppAuthentication/Services/UserManagementService/UserManagementService.cs
@@ -10,29 +10,27 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
     {
         private readonly RoleManager<IdentityRole> _roleManager;
         private readonly UserManager<ApplicationUser> _userManager;
-        private readonly ApplicationDbContext _context;
+        
         public UserManagementService (RoleManager<IdentityRole> roleManager, 
-                                      UserManager<ApplicationUser> userManager,
-                                      ApplicationDbContext context)
+                                      UserManager<ApplicationUser> userManager)
         {
             _roleManager = roleManager;
             _userManager = userManager;
-            _context = context;
         }
 
-        public List <IdentityRole> GetRoles () // все роли
+        public List <IdentityRole> GetRoles () 
         {
             return _roleManager.Roles.ToList();
         }
 
-        public async Task<string> GetRoleByUserAsync(ApplicationUser user) // роль для пользователя
+        public async Task<string> GetRoleByUserAsync(ApplicationUser user) 
         {
             var role = await _userManager.GetRolesAsync(user);
 
             return role?.FirstOrDefault()??string.Empty;
         }
 
-        public List<ApplicationUser> GetUsers() // всех юзеров
+        public List<ApplicationUser> GetUsers() 
         {
             return _userManager.Users?.ToList()??new List<ApplicationUser>();
         }
@@ -53,7 +51,7 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
             return users;
         }
 
-        public List<SelectListItem> GetSelectListRoles() // создание списка для выпадающего списка ролей
+        public List<SelectListItem> GetSelectListRoles() 
         {
             var selectListRole = new List<SelectListItem>();
             var roles = _roleManager.Roles.ToList();
@@ -68,7 +66,7 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
             return selectListRole;
         }
 
-        public List<SelectListItem> GetSelectListEmail() // создание списка для выпадающего списка emails
+        public List<SelectListItem> GetSelectListEmail() 
         {
             var selectListEmail = new List<SelectListItem>();
             var emails = _userManager.Users.Select(p => p.Email).ToList();
@@ -83,7 +81,7 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
             return selectListEmail;
         }
 
-        public List<SelectListItem> GetSelectListFirstName() // создание списка для выпадающего списка firstNames
+        public List<SelectListItem> GetSelectListFirstName() 
         {
             var selectListFirstName = new List<SelectListItem>();
             var firstNames = _userManager.Users.Select(p => p.FirstName).ToList();
@@ -100,7 +98,7 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
 
         public async Task<IdentityResult> CreateRoleAsync(string name)
         {
-            if (!string.IsNullOrEmpty(name) && !await IsRoleExiste(name))
+            if (!string.IsNullOrEmpty(name) && !await IsRoleExisteAsync(name))
             {
                 IdentityResult result = await _roleManager.CreateAsync(new IdentityRole(name));
                 
@@ -112,7 +110,7 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
             }
         }
 
-        public Task<bool> IsRoleExiste(string name)
+        public Task<bool> IsRoleExisteAsync(string name)
         {
             Task<bool> result = _roleManager.RoleExistsAsync(name);
             
@@ -121,10 +119,12 @@ namespace PizzaWebAppAuthentication.Services.RoleManagementService
 
         public async Task<IdentityResult> Delete (string name)
         {
-            if (!string.IsNullOrEmpty(name) && await IsRoleExiste(name))
+            if (!string.IsNullOrEmpty(name) && await IsRoleExisteAsync(name) && name != "Admin" && name != "User")
             {
-                IdentityRole? role = await _roleManager.FindByNameAsync(name);
-                if (role != null)
+                var role = await _roleManager.FindByNameAsync(name);
+                var isExistUser = await GetUsersByRoleAsync(name);
+
+                if (role != null && !isExistUser.Any())
                 {
                     IdentityResult result = await _roleManager.DeleteAsync(role);
